+++ tests/test_gather_skeleton_points.py
@@ -1,6 +1,12 @@
-"""Unit tests for `gather_skeleton_points()`."""
+# SPDX-FileCopyrightText: 2020-2023 CERN
+# SPDX-FileCopyrightText: 2023 GSI Helmholtzzentrum für Schwerionenforschung
+# SPDX-FileNotice: All rights not expressly granted are reserved.
+#
+# SPDX-License-Identifier: GPL-3.0-or-later OR EUPL-1.2+
 
 # pylint: disable = redefined-outer-name
+
+"""Unit tests for `gather_skeleton_points()`."""
 
 import typing as t
 from unittest.mock import Mock