+++ CHANGELOG.md
@@ -1,8 +1,16 @@
+<!--
+SPDX-FileCopyrightText: 2020-2023 CERN
+SPDX-FileCopyrightText: 2023 GSI Helmholtzzentrum für Schwerionenforschung
+SPDX-FileNotice: All rights not expressly granted are reserved.
+
+SPDX-License-Identifier: GPL-3.0-or-later OR EUPL-1.2+
+-->
+
 # Changelog
 
 ## Unreleased
 
-No changes yet!
+- OTHER: Open-source this package by adding the appropriate license notices.
 
 ## v0.2.0